fix: placeholder rewrite
