+++ src/components/Placeholder/__stories__/Block.stories.tsx
@@ -1,0 +1,7 @@
+import { Template } from './Template.stories'
+
+export const Block = Template.bind({})
+
+Block.args = {
+  variant: 'block',
+}
