+++ src/components/Placeholder/__stories__/Blocks.stories.tsx
@@ -1,0 +1,7 @@
+import { Template } from './Template.stories'
+
+export const Blocks = Template.bind({})
+
+Blocks.args = {
+  variant: 'blocks',
+}
