+++ src/components/Placeholder/__stories__/Box.stories.tsx
@@ -1,0 +1,8 @@
+import { Template } from './Template.stories'
+
+export const Box = Template.bind({})
+
+Box.args = {
+  length: 8,
+  variant: 'box',
+}
