+++ src/components/Placeholder/__stories__/Donut.stories.tsx
@@ -1,0 +1,7 @@
+import { Template } from './Template.stories'
+
+export const Donut = Template.bind({})
+
+Donut.args = {
+  variant: 'donut',
+}
