+++ src/components/Placeholder/__stories__/Line.stories.tsx
@@ -1,0 +1,7 @@
+import { Template } from './Template.stories'
+
+export const Line = Template.bind({})
+
+Line.args = {
+  variant: 'line',
+}
