+++ src/components/Placeholder/__stories__/List.stories.tsx
@@ -1,0 +1,8 @@
+import { Template } from './Template.stories'
+
+export const List = Template.bind({})
+
+List.args = {
+  length: 2,
+  variant: 'list',
+}
