+++ src/components/Placeholder/__stories__/Slider.stories.tsx
@@ -1,0 +1,7 @@
+import { Template } from './Template.stories'
+
+export const Slider = Template.bind({})
+
+Slider.args = {
+  variant: 'slider',
+}
