+++ src/components/Placeholder/__stories__/Template.stories.tsx
@@ -1,0 +1,6 @@
+import { ComponentStory } from '@storybook/react'
+import Placeholder from '..'
+
+export const Template: ComponentStory<typeof Placeholder> = ({ ...props }) => (
+  <Placeholder {...props} />
+)
